app/api/pyproject.toml
[project]
name = "mirai-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic[email]>=2.0.0",
    "websockets>=11.0.0",
    "requests>=2.31.0"
]
[project.optional-dependencies]
dev = ["pytest>=8.2.0","httpx>=0.27.0","ruff>=0.5.0","mypy>=1.10.0","python-jose>=3.3.0"]
[tool.ruff]
line-length = 120
[tool.ruff.lint]
select = ["E","F","I"]
[tool.setuptools.package-data]
"mirai_api" = ["py.typed"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true